+++ src/main/java/blackjack/view/OutputView.java
@@ -1,7 +1,6 @@
 package blackjack.view;
 
 import blackjack.domain.ParticipantResult;
-import blackjack.domain.ScoreCompareResult;
 import java.util.List;
 import java.util.Map;
 
@@ -45,10 +44,10 @@ public class OutputView {
         }
     }
 
-    public static void printFinalProfit(int dealerProfit, Map<String,Integer> playersProfit) {
+    public static void printFinalProfit(long dealerProfit, Map<String,Long> playersProfit) {
         System.out.println("\n## 최종 수익");
         System.out.println("딜러: " + dealerProfit);
-        for (Map.Entry<String, Integer> entry : playersProfit.entrySet()) {
+        for (Map.Entry<String, Long> entry : playersProfit.entrySet()) {
             System.out.println(entry.getKey() + ": " + entry.getValue());
         }
     }
