+++ src/test/java/blackjackTest/service/BlackjackTest.java
@@ -45,14 +45,14 @@ public class BlackjackTest {
         pobi.receiveOneCard(new Card(Rank.ACE, Shape.HEART));
         pobi.receiveOneCard(new Card(Rank.TEN, Shape.SPADE));
 
-        int profit = pobi.calculateFinalProfit(dealer);
-        int expectedAmount = (int) (bettingAmountValue * 1.5);
+        long profit = pobi.calculateFinalProfit(dealer);
+        long expectedAmount = (long) (bettingAmountValue * 1.5);
         assertThat(profit).isEqualTo(expectedAmount);
     }
 
     @Test
     void 플레이어_카드합_21_넘으면_배팅금액_모두_잃음() {
-        int bettingAmountValue = 10000;
+        long bettingAmountValue = 10000;
         Money bettingAmount = new Money(bettingAmountValue);
         Player pobi = new Player("pobi", bettingAmount);
         Dealer dealer = new Dealer();
@@ -61,14 +61,14 @@ public class BlackjackTest {
         pobi.receiveOneCard(new Card(Rank.TEN, Shape.SPADE));
         pobi.receiveOneCard(new Card(Rank.TWO, Shape.CLOVER));
 
-        int profit = pobi.calculateFinalProfit(dealer);
-        int expectedAmount = -bettingAmountValue;
+        long profit = pobi.calculateFinalProfit(dealer);
+        long expectedAmount = -bettingAmountValue;
         assertThat(profit).isEqualTo(expectedAmount);
     }
 
     @Test
     void 딜러_초기카드_합_21이면_배팅금액_모두_잃음() {
-        int bettingAmountValue = 10000;
+        long bettingAmountValue = 10000;
         Money bettingAmount = new Money(bettingAmountValue);
         Player pobi = new Player("pobi", bettingAmount);
         Dealer dealer = new Dealer();
@@ -78,14 +78,14 @@ public class BlackjackTest {
         dealer.receiveOneCard(new Card(Rank.ACE, Shape.CLOVER));
         dealer.receiveOneCard(new Card(Rank.QUEEN, Shape.SPADE));
 
-        int profit = pobi.calculateFinalProfit(dealer);
-        int expectedAmount = -bettingAmountValue;
+        long profit = pobi.calculateFinalProfit(dealer);
+        long expectedAmount = -bettingAmountValue;
         assertThat(profit).isEqualTo(expectedAmount);
     }
 
     @Test
     void 딜러와_플레이어_모두_초기카드_합_21이면_배팅금액_돌려받음() {
-        int bettingAmountValue = 10000;
+        long bettingAmountValue = 10000;
         Money bettingAmount = new Money(bettingAmountValue);
         Player pobi = new Player("pobi", bettingAmount);
         Dealer dealer = new Dealer();
@@ -95,13 +95,13 @@ public class BlackjackTest {
         dealer.receiveOneCard(new Card(Rank.ACE, Shape.CLOVER));
         dealer.receiveOneCard(new Card(Rank.QUEEN, Shape.SPADE));
 
-        int profit = pobi.calculateFinalProfit(dealer);
+        long profit = pobi.calculateFinalProfit(dealer);
         assertThat(profit).isEqualTo(bettingAmountValue);
     }
 
     @Test
     void 플레이어가_딜러보다_점수가_높으면_배팅금액_돌려받음() {
-        int bettingAmountValue = 10000;
+        long bettingAmountValue = 10000;
         Money bettingAmount = new Money(bettingAmountValue);
         Player pobi = new Player("pobi", bettingAmount);
         Dealer dealer = new Dealer();
@@ -113,7 +113,7 @@ public class BlackjackTest {
         dealer.receiveOneCard(new Card(Rank.EIGHT, Shape.CLOVER));
         dealer.receiveOneCard(new Card(Rank.QUEEN, Shape.SPADE));
 
-        int profit = pobi.calculateFinalProfit(dealer);
+        long profit = pobi.calculateFinalProfit(dealer);
 
         assertThat(profit).isEqualTo(bettingAmountValue);
     }
